refactor: made it responsive, added text in component
