+++ src/componentes/FavouriteDex/FavoriteDex.module.scss
@@ -14,11 +14,11 @@
   align-items: center;
 }
 
-.textContainer{
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-    align-items: start;
+.textContainer {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  align-items: start;
 }
 .separador {
   display: flex;
@@ -26,8 +26,8 @@
 }
 
 .subtext {
-    display: block;
-    margin:0;
+  display: block;
+  margin: 0;
 }
 .buttons {
   display: flex;
@@ -41,7 +41,7 @@
   height: 50px;
 }
 
-.imgContainer{
+.imgContainer {
   display: flex;
   overflow: hidden;
 }
@@ -50,43 +50,82 @@
   object-fit: cover;
 }
 
-@media (max-width:768px){
-  .favoriteDex{
+@media (max-width: 768px) {
+  .favoriteDex {
     flex-direction: column;
     align-items: center;
     justify-content: center;
   }
-  .container{
+  .container {
     display: flex;
     gap: 20px;
-
   }
-  .imgContainer{
+  .imgContainer {
     width: 500px;
     height: 500px;
   }
-  .textContainer{
+  .textContainer {
     flex-wrap: nowrap;
     align-items: center;
     flex-direction: column;
     max-width: 100%;
     object-fit: contain;
   }
-  .title{
+  .title {
     display: flex;
     flex-direction: row;
     gap: 20px;
   }
-  .separador{
+  .separador {
     gap: 20px;
   }
-  .subtext{
+  .subtext {
     display: flex;
     align-items: center;
     justify-content: center;
     flex-wrap: nowrap;
   }
-  .buttons{
+  .buttons {
+    justify-content: center;
+  }
+}
+
+@media (max-width: 425px) {
+  .favoriteDex {
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+  .container {
+    display: flex;
+    gap: 20px;
+  }
+  .imgContainer {
+    width: 400px;
+    height: 400px;
+  }
+  .textContainer {
+    flex-wrap: nowrap;
+    align-items: center;
+    flex-direction: column;
+    max-width: 100%;
+    object-fit: contain;
+  }
+  .title {
+    display: flex;
+    flex-direction: row;
+    gap: 20px;
+  }
+  .separador {
+    gap: 20px;
+  }
+  .subtext {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-wrap: wrap;
+  }
+  .buttons {
     justify-content: center;
   }
 }
